+++ app/policies/guild_invite_policy.rb
@@ -1,7 +1,13 @@
 # GuildInvite policies
 class GuildInvitePolicy < ApplicationPolicy
+  authorize :guild, :character
+
   def new?
-    return user.any_role?(record.id, 'gm', 'rl') if record.is_a?(Guild)
-    true
+    return user.any_role?(guild.id, 'gm', 'rl') if record == 'true'
+    character.user_id == user.id && character.guild_id.nil?
+  end
+
+  def create?
+    new?
   end
 end