+++ app/serializers/guild_invite_serializer.rb
@@ -1,0 +1,3 @@
+class GuildInviteSerializer < ActiveModel::Serializer
+  attributes :id, :guild_id, :character_id, :from_guild
+end