+++ frontend/components/guild_invite_new/guild_invite_new.js
@@ -1,0 +1,1 @@
+import "./guild_invite_new.scss";