+++ frontend/packs/invite_form_for_guild.jsx
@@ -1,0 +1,4 @@
+import InviteFormForGuild from "components_react/invite_form_for_guild/invite_form_for_guild";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({InviteFormForGuild});