+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - left guild by characters
 - generate new gm or disbund the guild
 - GuildInvite model
-- page for invites
+- page for guild invites
+- api endpoint for creating guild invites
 
 ## [0.6.4] - 2019-11-06
 ### Added